+++ js/server.js
@@ -3,14 +3,14 @@ import {showAlert} from './util.js';
 import {renderModal, toDefaultForm} from './form.js';
 
 
-function getData() {
-  fetch('https://22.javascript.pages.academy/keksobooking/data')
-    .then((response) => response.json())
-    .then((data) => renderToMap(data.slice(0, 10)))
-    .catch(() => {
-      showAlert('При загрузке данных с сервера произошла ошибка. Попробуйте ещё раз');
-    });
-}
+// function getData() {
+//   fetch('https://22.javascript.pages.academy/keksobooking/data')
+//     .then((response) => response.json())
+//     .then((data) => renderToMap(data.slice(0, 10)))
+//     .catch(() => {
+//       showAlert('При загрузке данных с сервера произошла ошибка. Попробуйте ещё раз');
+//     });
+// }
 
 function toSend(data) {
   fetch(
@@ -33,6 +33,6 @@ function toSend(data) {
 }
 
 export {
-  getData,
+  // getData,
   toSend
 };
